backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
arq==0.27.0
asyncpg==0.31.0
certifi==2026.2.25
click==8.3.2
coverage==7.13.5
distro==1.9.0
docstring-parser==0.18.0
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
hiredis==3.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.14.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.31.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
PyYAML==6.0.3
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-cov==7.1.0
python-dotenv==1.2.2
python-multipart==0.0.26
redis==5.3.1
sniffio==1.3.1
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
